requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.1.6
django-filter==25.1
django-phonenumber-field==8.0.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
idna==3.10
phonenumbers==8.13.55
pillow==11.1.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
requests==2.32.3
rest-framework-simplejwt==0.0.2
sqlparse==0.5.3
typing_extensions==4.12.2
urllib3==2.3.0
